third_party/WebKit/LayoutTests/html/document-metadata/base-expected.txt

PASS endsWith(document.querySelector('base').href, 'foo/bar/') is true
PASS endsWith(document.querySelector('base').href, 'foo/bar/foo/bar/') is false
PASS endsWith(document.querySelector('base').href, '/null') is true
PASS documentWithoutAView.querySelector('base').href is ""
PASS documentWithoutAView.querySelector('base').href is "http://webkit.org/"
PASS iframe.contentDocument.querySelector('base').href is ""
PASS newWindow.document.querySelector('base').href is ""
PASS successfullyParsed is true

TEST COMPLETE
